Enable NetworkPolicy in the helm chart
We enable NetworkPolicy in the z2jh helm chart
(http://z2jh.jupyter.org/en/latest/security.html#kubernetes-network-policies).

This provides basic firewalling for the various components we run,
making sure they are only making the network requests they are
supposed to make.

1. Hub can only talk to the proxy and user pods
2. Proxy can only talk to the hub and user pods
3. User pods can only talk to hub, proxy and port 53/80/443 to
   everything else.

We might need to relax these permissions later for other
reasons, but it's a good start.

The cluster on which we currently run most on-campus hubs and
data8x hub don't have NetworkPolicy enforcement (https://cloud.google.com/kubernetes-engine/docs/how-to/network-policy)
enabled, so this should be a no-op there.  The 'summer-2019' cluster
created for hubs this summer *does* have it enabled, so we can
migrate the hubs and make sure they are working over the summer.