Add a user placeholder rebalancer
When a cluster scales up, user pods are scheduled on to it immediately.
The user images aren't present on this node, and our images are so
big it takes about 10min for the images to land there. During this
time, all user pods scheduled on the new node fail, since the timeout
for new user spawns is 5min. The upshot is that whenever a new
node comes up, user spawns *fail* for about 10 minutes. This is
bad.

We have user placeholder pods that trigger scale-ups when the
cluster is about 90% full - this means new nodes come up before
the cluster is 100% full. However, because the new node is emptier
than the older nodes, user pods get scheduled on the new node -
even though it doesn't have the user images on it!

This adds a rebalancer script, that runs every 10s and does the
following:

1. Detects if a new node has been added
2. If so, it labels it a particular way that is set to attract the
   placeholder pods
3. It deletes all the placeholder pods, so they end up on the new node.

This frees up enough space in the older nodes, so user pods end up
there instead of in the newer, imageless node.

This is a stop-gap solution, and there could be many improvements
made to it - ultimately it will be superseeded by a Kubernetes
Feature directly. https://github.com/jupyterhub/zero-to-jupyterhub-k8s/issues/1414
has more information & pointers.

We also automate the building of the hub & rebalancer image, so
that no longer needs to be done manually