Remove chartpress from CD
Currently, chartpress affects deployment of each hub, and
thus blocks them from being done separately. Ideally, we'd
be able to build the images in one step, and then just use the tags
in another step. chartpress currently does not support this,
so we just remove it from CD. We'll have to run it manually
for now.