datahub: Upgrade all packages to latest versions
Unfortunately, this seems to be the only way to do package
bumps right now - we can't really do them in the middle of
the semester. A lot of packages don't work with Python 3.7,
or don't have manylinux wheels.

I'll actively reach out to folks, and we can revert individual
libraries that cause issues.