Delete rstudio state file on startup
Seems to be causing a bunch of blocking issues for students,
where they are unable to access the R hub because of possible
corruption in RStudio. Deleting / renaming the state file
seems to do the trick, but doing that on a per-student basis -
especially if they are blocked from using the hub in the meantime -
seems bad.

This might affect classes that rely on persistent R state,
but in general I think that's not a recommended pattern -
preferring state to be in explicitly saved files &
code itself (I could be totally wrong!). We can merge this
to unblock users, and see if others complain about the lack
of state?

Ref #1899