cs194: Provide persistent storage for postgres
For each user, we create a PVC and mount the postgres
database directory there. We also mount the same directory
in the user pod, in case direct inspection is necessary.

Each user gets 1G ssd. For about 200 users, that's 200G ssd -
which is as much each node gets. I think we're ok cost-wise for now.

Ref #2116