Add node placeholders
We currently use 'pod placeholders'
(https://zero-to-jupyterhub.readthedocs.io/en/stable/administrator/optimization.html)
to try keep 'headroom' of X nodes over what users are currently
using - so when a new user comes in, they can get a node right
away without needing to wait for a spinup. In practice, this has
been problematic - primarily because the unit of placeholder is *pod*,
but the unit of autoscaler is *node*. A common failure mode is

1. A new user comes in, displayes a single placeholder pod. This kicks
   off the autoscaler, and a new node spins up.
2. Another user comes in - but since the new node is emptier than the
   old node, the new user is scheduled in the new node! They have to
   then wait for the image to be pulled, which can take a long time.

So users still see a long wait, even though there's an 'empty'
node.

With the node placeholder approach, we make one pod that tries
to take up *all* the resources on a node. When a user pod comes up,
it can displace this *one* pod - and free up a full node of resources!
The displayed pod will trigger a scale-up, but since not even a single
other user pod can be on the same node as our placeholder pod, the new
user pods will continue to be on the recently vacated node.

So in our case, I've added enough headroom for 3 extra delta
nodes - primarily for use by the EECS hub. This should theoretically
help with a surge of upto 150 users over the course of new node spinup
+ ready - about 15 minutes or so.

Let's try it out