Specify TZ explicitly for RStudio
RStudio doesn't load environment variables in the
container (grr), so we need to explicitly set them again.
Looks like something is intermittently trying to get the
current TZ by calling `timedatectl`, which is a systemd
thing. We don't run systemd in this container, so this
fails / provides a scary warning. A bit hard to reproduce,
but hopefully this will help?