Setup prometheus_statsd to collect popularity contest metrics
We want to collect aggregated information about which
packages are being used, but we do not want to identify
individual users. statsd is perfect for it - we can push
data to it when the process exits, and it globally
aggregates our metrics.

We parse out useful labels from statsd metric names,
so we can run useful queries over it in prometheus