Put data8 in its own pool
The alpha-pool has datahub in it, with its huge image. New
node spinup takes up extra time in the alpha pool, causing delays
in how long it takes data8 spikes to be handled. With a special
pool just for data8, we can reduce this spin-up time by only keeping
the data8 image on it.

The data8 pool also has SSD disks for faster image pulls - we can
reduce that later if necessary, but I think this should help with
image pull speeds.