Disable user scheduler for everything except data8x
Trying out the optimize-utilization autoscaler profile
in
https://cloud.google.com/kubernetes-engine/docs/concepts/cluster-autoscaler#autoscaling_profiles,
which comes with its own scheduler. Hopefully it can replicate the
behavior we get with user-scheduler, and we can reduce the amount
of resources we spend on core nodes. More importantly, this will
reduce race conditions between the default scheduler (used for the
node placeholders) and the user scheduler (used for pods). It was
causing 'node does not have enough memory' after the pod was
assigned to the node, so those nodes then just get stuck in limbo
forever.