Move everything except data8 to alpha-pool
With the dynamic node placeholder scaler, we can automatically
increase the size of the cluster just before big classes. The
gamma node pool was helpful in rapid scale up, as the node pools
can be scaled independently, and not all images needed to be pulled
into all the nodes. However, two node pools lead to lower utilization
when some pods are spread out across the nodes in the node pool - we
end up with 6 nodes where we could have 3.

The new alpha node pool has SSD disks, and hopefully can come up
fast enough (when used with the placeholder) for all hubs except
data8.

Ref https://github.com/berkeley-dsep-infra/datahub/issues/2672