Get some python packages from conda-forge
Cartopy and Shapely feel very difficult to install from PyPI,
as they have no binary wheels and require a specific version of
proj and geos from apt. But the R `sf` package also requires
a specific version of proj from apt, and this sort of puts us
in an impossible situation as cartopy 0.20.* requires a newer
version than sf.

So we start using conda-forge to install some of these base
scientific packages instead. This reduces the scope for a lot
of these conflicts, and lets us peacefully use apt just for R.
I'm trying to only have pip packages depend on conda, not
the other way around, so I've moved some of the 'base'
scientific python packages to be installed from conda.
Ideally, we should basically move everything.

Ref https://github.com/berkeley-dsep-infra/datahub/issues/2824