Bust CircleCI cache
CircleCI is erroring with 'gcloud not found', let's see
iff this fixes it?

We updated the version of gcloud in an earlier PR, but did
not update the string used to keep the cache of our installed
python packages. I'm wondering if that is what is causing
the gcloud command to be not found in path