Limit biology hub users to 4 CPUs
A couple of biology hub users have been saturating the CPU on
the nodes they're on, impacting performance on other hubs in
the beta nodepool (ischool and public health). Bringing them
down to a limit of 4 CPUs should help. If this continues, we can
just move biology to its own nodepool.